requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0
assemblyai==0.33.0
google-generativeai==0.3.2
pydantic==2.5.0